initial structure new components
