+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "zustand": "^4.4.1"
+    "zustand": "^4.4.4"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
