+++ src/components/CounterComponent/CounterComponent.module.css
@@ -1,0 +1,1 @@
+/* CSS for the counter component component, put styling here */
