+++ src/components/Header/Header.jsx
@@ -1,0 +1,9 @@
+import styles from "./Header.module.css";
+
+export const Header = () => {
+  return (
+    <div className={styles.header}>
+      <h1>Header component - Country Quiz!</h1>
+    </div>
+  );
+};
