+++ src/components/Header/Header.module.css
@@ -1,0 +1,5 @@
+/* CSS for the header component, put styling here */
+
+.header {
+  color: blue;
+}
